test bug str equality
