+++ tests/test_file_processing.py
@@ -85,6 +85,7 @@ from crawler.constants import (
 from crawler.db.mongo import get_mongo_collection
 from crawler.file_processing import ERRORS_DIR, SUCCESSES_DIR, Centre, CentreFile
 from crawler.types import Config, ModifiedRow
+from tests.conftest import generate_new_object_for_string
 
 # ----- tests helpers -----
 
@@ -1571,7 +1572,7 @@ def test_insert_plates_and_wells_from_docs_into_dart_multiple_new_plates(config)
             FIELD_PLATE_BARCODE: "TC-rna-00000029",
             FIELD_COORDINATE: "A01",
             FIELD_LAB_ID: "AP",
-            FIELD_RESULT: POSITIVE_RESULT_VALUE,
+            FIELD_RESULT: generate_new_object_for_string(POSITIVE_RESULT_VALUE),
         },
         {
             "_id": ObjectId("5f562d9931d9959b92544728"),
@@ -1580,7 +1581,7 @@ def test_insert_plates_and_wells_from_docs_into_dart_multiple_new_plates(config)
             FIELD_PLATE_BARCODE: "TC-rna-00000024",
             FIELD_COORDINATE: "B01",
             FIELD_LAB_ID: "AP",
-            FIELD_RESULT: POSITIVE_RESULT_VALUE,
+            FIELD_RESULT: generate_new_object_for_string(POSITIVE_RESULT_VALUE),
         },
         {
             "_id": ObjectId("5f562d9931d9959b92544728"),
@@ -1650,7 +1651,7 @@ def test_insert_plates_and_wells_from_docs_into_dart_single_new_plate_multiple_w
             FIELD_PLATE_BARCODE: plate_barcode,
             FIELD_COORDINATE: "A01",
             FIELD_LAB_ID: "AP",
-            FIELD_RESULT: POSITIVE_RESULT_VALUE,
+            FIELD_RESULT: generate_new_object_for_string(POSITIVE_RESULT_VALUE),
         },
         {
             "_id": ObjectId("5f562d9931d9959b92544728"),
@@ -1659,7 +1660,7 @@ def test_insert_plates_and_wells_from_docs_into_dart_single_new_plate_multiple_w
             FIELD_PLATE_BARCODE: plate_barcode,
             FIELD_COORDINATE: "A02",
             FIELD_LAB_ID: "AP",
-            FIELD_RESULT: POSITIVE_RESULT_VALUE,
+            FIELD_RESULT: generate_new_object_for_string(POSITIVE_RESULT_VALUE),
         },
         {
             "_id": ObjectId("5f562d9931d9959b92544728"),
